requirements.txt
Django
openai
google-genai